MAINT: Move adjustments module.
Incoming changes write more than pipeline related data to the
adjustments module, so move module ahead of time to make the diff more
legible.
